Multiple test fixes and refactors
