+++ tests/src/rpc/client.spec.ts
@@ -3,10 +3,9 @@
  * @description Unit test for the Pocket Core Client interface
  */
 import { expect } from "chai"
-import { Pocket, BondStatus, Configuration, HttpRpcProvider, 
-    Node, PocketAAT, EnvironmentHelper, Network, RpcError, typeGuard, RelayResponse
-} from "../../../src"
 import { NockUtil } from "../../utils/nock-util"
+import { EnvironmentHelper } from "../../../src/utils/env/helper"
+import { Node, PocketAAT, BondStatus, Configuration, HttpRpcProvider, Pocket, typeGuard, RpcError } from "../../../src"
 
 // Constants
 // For Testing we are using dummy data, none of the following information is real.
@@ -36,13 +35,75 @@ const noSessionPocketAAT = PocketAAT.from(version, clientPublicKey, alternatePub
 const node01 = new Node(addressHex, applicationPublicKey, false, BondStatus.bonded, BigInt(100), env.getPOKTRPC(), blockchains)
 const configuration = new Configuration([node01], 5, 40000, 200)
 const rpcProvider = new HttpRpcProvider(new URL(node01.serviceURL))
+const nodeAddress = "84871BAF5B4E01BE52E5007EACF7048F24BF57E0"
+const nodePublicKey = "d9c7f275388ca1f87900945dba7f3a90fa9bba78f158c070aa12e3eccf53a2eb"
+const testNode = new Node(nodeAddress, nodePublicKey, false, BondStatus.bonded, BigInt(100), env.getPOKTRPC(), ["ETH04"])
+
 
 // Default pocket instance to reuse code
 function getPocketDefaultInstance(): Pocket {
     return new Pocket(configuration, rpcProvider)
 }
 
+function defaultConfiguration(): Configuration {
+    return new Configuration([testNode])
+}
+
+function createPocketInstance(configuration?: Configuration): Pocket {
+    if (configuration === undefined) {
+        const baseURL = new URL(defaultConfiguration().nodes[0].serviceURL)
+        const rpcProvider = new HttpRpcProvider(baseURL)
+        return new Pocket(defaultConfiguration(), rpcProvider)
+    } else {
+        const baseURL = new URL(configuration.nodes[0].serviceURL)
+        const rpcProvider = new HttpRpcProvider(baseURL)
+        return new Pocket(configuration, rpcProvider)
+    }
+}
+
 describe("Pocket RPC Client Interface", async () => {
+    describe("Sending raw transactions", function () {
+        describe("Success scenarios", function () {
+            it("should succesfully send a raw transaction given a fromAddress, tx hex and Node", async () => {
+                NockUtil.mockRawTx()
+                const pocket = createPocketInstance()
+                const senderAddressHex = "11AD05777C30F529C3FD3753AD5D0EA97192716E"
+                const txHex = "c001db0b170d0a3c939866b00a1411ad05777c30f529c3fd3753ad5d0ea97192716e12149e8e373ff27ec202f82d07df64f388ff42f9516d1a0a0a04706f6b7412023130120b0a04706f6b7412033130301a690a259d54477420917fe8e7fc02ceabddfbb10168dcd6885180d9f2db8855dbe063f6c5f7f93c9c12402c40788aac1e27539647a14a447d89966ab5fa6d4380c2b24a839c91be0106fbae55b76a9168a780108a40466b85b190ef12b9b6db6a879f6daa3b70b090100a22046c6f6c6f"
+                const response = await pocket.rpc.client.rawtx(senderAddressHex, txHex, pocket.configuration.requestTimeOut)
+                expect(response).not.to.be.a('error')
+            })
+        })
+
+        describe("Error scenarios", function () {
+            it("should fail given an invalid or empty address hex", async () => {
+                NockUtil.mockRawTx()
+
+                // Invalid address
+                const pocket = createPocketInstance()
+                let senderAddressHex = "11AD05777C30F529C3FD3753AD5D0EA9719271"
+                let txHex = "c001db0b170d0a3c939866b00a1411ad05777c30f529c3fd3753ad5d0ea97192716e12149e8e373ff27ec202f82d07df64f388ff42f9516d1a0a0a04706f6b7412023130120b0a04706f6b7412033130301a690a259d54477420917fe8e7fc02ceabddfbb10168dcd6885180d9f2db8855dbe063f6c5f7f93c9c12402c40788aac1e27539647a14a447d89966ab5fa6d4380c2b24a839c91be0106fbae55b76a9168a780108a40466b85b190ef12b9b6db6a879f6daa3b70b090100a22046c6f6c6f"
+                const invalidResponse = await pocket.rpc.client.rawtx(senderAddressHex, txHex, pocket.configuration.requestTimeOut)
+                expect(typeGuard(invalidResponse, RpcError)).to.be.true
+
+                // Empty address
+                senderAddressHex = ""
+                txHex = "c001db0b170d0a3c939866b00a1411ad05777c30f529c3fd3753ad5d0ea97192716e12149e8e373ff27ec202f82d07df64f388ff42f9516d1a0a0a04706f6b7412023130120b0a04706f6b7412033130301a690a259d54477420917fe8e7fc02ceabddfbb10168dcd6885180d9f2db8855dbe063f6c5f7f93c9c12402c40788aac1e27539647a14a447d89966ab5fa6d4380c2b24a839c91be0106fbae55b76a9168a780108a40466b85b190ef12b9b6db6a879f6daa3b70b090100a22046c6f6c6f"
+                const emptyResponse = await pocket.rpc.client.rawtx(senderAddressHex, txHex, pocket.configuration.requestTimeOut)
+                expect(typeGuard(emptyResponse, RpcError)).to.be.true
+            })
+
+            it("should fail given an empty tx hex", async () => {
+                NockUtil.mockRawTx()
+                // Empty address
+                const pocket = createPocketInstance()
+                const senderAddressHex = "11AD05777C30F529C3FD3753AD5D0EA9719271"
+                const txHex = ""
+                const response = await pocket.rpc.client.rawtx(senderAddressHex, txHex, pocket.configuration.requestTimeOut)
+                expect(typeGuard(response, RpcError)).to.be.true
+            })
+        })
+    })
+
     describe("Success scenarios", async () => {
          // TODO: Fix signing
         //  it('should successfully send a relay due to a valid information', async () => {
